refactor(program-logs): move event extraction into the entity
