+++ app/entities/program-logs/index.ts
@@ -3,3 +3,4 @@ export { TxExecutionStatus } from './ui/TxExecutionStatus';
 export { TxSimulationStatus } from './ui/TxSimulationStatus';
 export { TxErrorStatus } from './ui/TxErrorStatus';
 export { useParsedLogs } from './model/useParsedLogs';
+export { extractEventsFromLogs, type ProgramEventPayload } from './model/extract-events';
